feat: Add lazy loading for images and update createAsset page
This commit adds lazy loading for images in the application to improve page load performance. It also updates the "createAsset" page to include the functionality of creating an NFT using the "CreateNFT" function from the "blockchainFunctions/writeTx" module. The owner address is now obtained from the active account using the "useActiveAccount" hook. These changes are implemented in the file "app/[locale]/createProduct/page.jsx".